+++ src/Combinator/ICombination.cs
@@ -5,7 +5,7 @@ namespace Combinator
     public interface ICombination<T>
     {
         ImmutableList<T> Nodes { get; }
-        int Value { get; }
-        int Cost { get; }
+        double Value { get; }
+        double Cost { get; }
     }
 }
